#9 UI improvements done
