+++ client/src/components/Search/SearchCriteria.jsx
@@ -1,0 +1,39 @@
+import { useState } from "react";
+
+import { ButtonToolbar, Button } from "react-bootstrap";
+
+const SearchCriteria = () => {
+
+    const [activeButton, setActiveButton] = useState("name"); // Make the first button active initially
+
+
+    const buttons = [
+        { value: "name", text: "Name" },
+        { value: "specialization", text: "Specialization" },
+        { value: "designation", text: "Designation" },
+        { value: "currentWorkplace", text: "Current Workplace" },
+        { value: "institute", text: "Institute" },
+      ];
+    
+      const handleButtonClick = (buttonNumber) => {
+        setActiveButton(buttonNumber);
+      };
+
+      return (
+        <ButtonToolbar className="justify-content-between">
+        {buttons.map((button) => (
+          <Button
+            key={button.value}
+            variant={
+              activeButton === button.value ? "primary" : "secondary"
+            }
+            onClick={() => handleButtonClick(button.value)}
+          >
+            {button.text}
+          </Button>
+        ))}
+      </ButtonToolbar>
+      )
+}
+
+export default SearchCriteria;
