+++ client/src/schema/searchResultsTypes.ts
@@ -1,0 +1,8 @@
+export const SearchResultsTitles = {
+    designation: 'Designation',
+    workplace:'Current Workplace',
+    personalEmail:'Personal Email',
+    personalContactNumber:'Personal Contact Number',
+    officialEmail:'Official Email',
+}
+
